+++ src/main/java/pl/jkkk/cps/logic/model/window/Window.java
@@ -1,5 +1,5 @@
 package pl.jkkk.cps.logic.model.window;
 
 public interface Window {
-    double w(double n);
+    double w(int n);
 }
